requirements.txt
anthropic
openai-whisper
yt-dlp
fastapi
uvicorn[standard]
